backend/requirements.txt
django
djangorestframework
django-cors-headers
google-cloud-firestore
firebase-admin
python-dotenv
pillow
